os path fixes

+++ src/pcgr/pcgr_check_input.py
@@ -46,8 +46,10 @@ def is_valid_cna_segment_file(cna_segment_file, logger):
    if not ('Chromosome' in cna_reader.fieldnames and 'Segment_Mean' in cna_reader.fieldnames and 'Start' in cna_reader.fieldnames and 'End' in cna_reader.fieldnames):
       error_message_cnv = "Copy number segment file (" + str(cna_segment_file) + ") is missing required column(s): 'Chromosome', 'Start', 'End', or  'Segment_Mean'"
       error_message_cnv2 = "Column names present in file: " + str(cna_reader.fieldnames)
+      logger.error('')
       logger.error(error_message_cnv)
       logger.error(error_message_cnv2)
+      logger.error('')
       return -1
    
    cna_dataframe = np.read_csv(cna_segment_file, sep="\t")
@@ -69,10 +71,14 @@ def is_valid_cna_segment_file(cna_segment_file, logger):
 
    for rec in cna_reader:
       if int(rec['End']) < int(rec['Start']):
+         logger.error('')
          logger.error('Detected wrongly formatted chromosomal segment - \'Start\' is greater than \'End\' (' + str(rec['Chromosome']) + ':' + str(rec['Start']) + '-' + str(rec['End']) + ')')
+         logger.error('')
          return -1
       if rec['End'] < 1 or rec['Start'] < 1:
+         logger.error('')
          logger.error('Detected wrongly formatted chromosomal segment - \'Start\' or \'End\' is less than or equal to zero (' + str(rec['Chromosome']) + ':' + str(rec['Start']) + '-' + str(rec['End']) + ')')
+         logger.error('')
          return -1
    logger.info('Copy number segment file (' + str(cna_segment_file) + ') adheres to the correct format')
    return 0
@@ -82,21 +88,22 @@ def is_valid_cna_segment_file(cna_segment_file, logger):
 def is_valid_vcf(vcf_validator_output_file):
    
    valid_vcf = -1
-   f = open(vcf_validator_output_file,'r')
-   error_messages = []
-   for line in f:
-      if not re.search(r' \(warning\)$|^Reading from ',line.rstrip()):
-         if line.startswith('Line '):
-            error_messages.append(line.rstrip())
-         if line.endswith('the input file is valid'):
-            valid_vcf = 1
-         if line.endswith('the input file is not valid'):
-            valid_vcf = 0
-   f.close()
-   os.system('rm -f ' + str(vcf_validator_output_file))
    ret = {}
-   ret['error_messages'] = error_messages
-   ret['validation_status'] = valid_vcf
+   if os.path.exists(vcf_validator_output_file):
+      f = open(vcf_validator_output_file,'r')
+      error_messages = []
+      for line in f:
+         if not re.search(r' \(warning\)$|^Reading from ',line.rstrip()):
+            if line.startswith('Line '):
+               error_messages.append(line.rstrip())
+            if line.endswith('the input file is valid'):
+               valid_vcf = 1
+            if line.endswith('the input file is not valid'):
+               valid_vcf = 0
+      f.close()
+      os.system('rm -f ' + str(vcf_validator_output_file))
+      ret['error_messages'] = error_messages
+      ret['validation_status'] = valid_vcf
    return ret
 
 
@@ -135,18 +142,22 @@ def verify_input(input_vcf, input_cna_segments):
             POS = rec.start + 1
             alt = ",".join(str(n) for n in rec.ALT)
             if len(rec.ALT) > 1:
-               error_message_multiallelic = "Multiallelic site detected:" + str(rec.CHROM) + '\t' + str(POS) + '\t' + str(rec.REF) + '\t' + str(alt)
-               logger.error(error_message_multiallelic)
+               logger.error('')
+               logger.error("Multiallelic site detected:" + str(rec.CHROM) + '\t' + str(POS) + '\t' + str(rec.REF) + '\t' + str(alt))
                logger.error('Alternative alleles must be decomposed and normalized, see http://pcgr.readthedocs.io/en/latest/output.html#vcf-preprocessing')
+               logger.error('')
                multiallelic_alt = 1
                return -1
-         command_vcf_sample_free1 = 'egrep \'^#\' ' + str(input_vcf) + ' > ' + str(input_vcf_pcgr_ready)
-         command_vcf_sample_free2 = 'egrep -v \'^#\' ' + str(input_vcf) + ' | cut -f1-8 > ' + str(input_vcf_pcgr_ready)
+         command_vcf_sample_free1 = 'egrep \'^##\' ' + str(input_vcf) + ' > ' + str(input_vcf_pcgr_ready)
+         command_vcf_sample_free2 = 'egrep \'^#CHROM\' ' + str(input_vcf) + ' | cut -f1-8 >> ' + str(input_vcf_pcgr_ready)
+         command_vcf_sample_free3 = 'egrep -v \'^#\' ' + str(input_vcf) + ' | cut -f1-8 >> ' + str(input_vcf_pcgr_ready)
          if input_vcf.endswith('.gz'):
-            command_vcf_sample_free1 = 'bgzip -dc ' + str(input_vcf) + ' | egrep \'^#\' > ' + str(input_vcf_pcgr_ready)
-            command_vcf_sample_free2 = 'bgzip -dc ' + str(input_vcf) + ' | egrep -v \'^#\' | cut -f1-8 > ' + str(input_vcf_pcgr_ready)
+            command_vcf_sample_free1 = 'bgzip -dc ' + str(input_vcf) + ' | egrep \'^##\' > ' + str(input_vcf_pcgr_ready)
+            command_vcf_sample_free2 = 'bgzip -dc ' + str(input_vcf) + ' | egrep \'^#CHROM\' | cut -f1-8 >> ' + str(input_vcf_pcgr_ready)
+            command_vcf_sample_free3 = 'bgzip -dc ' + str(input_vcf) + ' | egrep -v \'^#\' | cut -f1-8 >> ' + str(input_vcf_pcgr_ready)
          os.system(command_vcf_sample_free1)
          os.system(command_vcf_sample_free2)
+         os.system(command_vcf_sample_free3)
          os.system('bgzip -f ' + str(input_vcf_pcgr_ready))
          os.system('tabix -p vcf ' + str(input_vcf_pcgr_ready) + '.gz')
